requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt
#
aiohttp==3.8.3
    # via cloud-sql-python-connector
aiosignal==1.3.1
    # via aiohttp
asn1crypto==1.5.1
    # via scramp
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
cachecontrol==0.12.11
    # via firebase-admin
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
cloud-sql-python-connector[pg8000]==1.1.0
    # via -r requirements.in
coverage==7.0.5
    # via -r requirements.in
cryptography==39.0.0
    # via
    #   cloud-sql-python-connector
    #   pyjwt
distlib==0.3.6
    # via virtualenv
filelock==3.9.0
    # via virtualenv
firebase-admin==6.1.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.11.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.79.0
    # via firebase-admin
google-auth==2.16.0
    # via
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.10.0
    # via firebase-admin
google-cloud-storage==2.7.0
    # via firebase-admin
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.1
    # via google-cloud-storage
googleapis-common-protos==1.58.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==2.0.1
    # via sqlalchemy
grpcio==1.51.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpcio-status
    #   grpcio-tools
grpcio-status==1.51.3
    # via google-api-core
grpcio-tools==1.51.1
    # via -r requirements.in
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
identify==2.5.13
    # via pre-commit
idna==3.4
    # via
    #   requests
    #   yarl
mock==5.0.1
    # via -r requirements.in
msgpack==1.0.4
    # via cachecontrol
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nodeenv==1.7.0
    # via pre-commit
pg8000==1.29.4
    # via cloud-sql-python-connector
platformdirs==2.6.2
    # via virtualenv
pre-commit==2.21.0
    # via -r requirements.in
proto-plus==1.22.2
    # via google-cloud-firestore
protobuf==4.21.12
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   grpcio-tools
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via firebase-admin
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via pg8000
python-dotenv==0.21.1
    # via -r requirements.in
pyyaml==6.0
    # via pre-commit
requests==2.28.2
    # via
    #   cachecontrol
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
scramp==1.4.4
    # via pg8000
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
sqlalchemy==1.4.46
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.14
    # via requests
virtualenv==20.17.1
    # via pre-commit
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
